adds wip restaurant search request
