+++ app/services/eatstreet/restaurant.rb
@@ -1,0 +1,11 @@
+module Eatstreet
+	class Restaurant < Base
+		attr_accessor :name, :streetAddress, :city, :foodTypes, :offersDelivery
+
+		# MAX_LIMIT = 10
+
+		def self.search(query = {}, options)
+			response = Request.where('restaurant/search', query)
+		end
+	end
+end
