+++ lib/eat_street/connection.rb
